% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{case_plot}
\alias{case_plot}
\title{Case plot for figures}
\usage{
case_plot(df, testing = FALSE, facet = "wrap", nrow = NULL, gridvar = NULL)
}
\arguments{
\item{df}{data frame of cases. Can be generated by \code{outbreak_model} (with \code{weekly = FALSE})}

\item{testing}{logical - do you want to plot testing rates?}

\item{facet}{character - 'wrap' if you want to plot one variable, 'grid' if you want to plot two}

\item{nrow}{integer - number of rows (only use if facet == 'wrap')}

\item{gridvar}{name of second variable if plotting two (first is always 'intervention')}
}
\value{

}
\description{
Case plot for figures
}
\examples{

}
\author{
Lewis Spurgin
}
