% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dist2b_func}
\alias{dist2b_func}
\title{Simulate social distancing using data across days no reassignment}
\usage{
dist2b_func(am, am.contacts, returns, dist.prop)
}
\description{
Simulate social distancing using data across days no reassignment
}
