% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dist_no_reall}
\alias{dist_no_reall}
\title{Simulate social distancing no reassignment}
\usage{
dist_no_reall(am, returns, dist.prop)
}
\arguments{
\item{am}{association matrix network}

\item{returns}{format for output - either "graph", "matrix" or "edgelist"}

\item{dist.prop}{the proportion of 'rare' ties to reassign. Between zero and one.}
}
\value{

}
\description{
Simulate social distancing no reassignment
}
\examples{

}
\author{
Josh Firth
}
