% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{draw.contagion}
\alias{draw.contagion}
\title{Generate network plot with contagion}
\usage{
draw.contagion(am, am.layout = NULL, use.df, day)
}
\arguments{
\item{am}{association matrix}

\item{am.layout}{(if wanting to specify a layout e.g. to keep the layout the same across comparisons)}

\item{use.df}{the dataframe information of the contagion}

\item{day}{the day to do the contagion for}
}
\value{

}
\description{
Generate network plot with contagion
}
\examples{

\dontrun{

#example of using function
a<-load(file="data-raw/am_list.RData") #need this for a network
b<-load(file="/network_examples.RData") #need this for contagion data


#here, we're going to draw 2 networks, both of the same day (time period of the contagion) and both in the same layout (but  assuming you don't have a pre-specified layout) then just like this works:
dev.new(height=6,width=12);par(mar=c(0,0,0,0));par(mai=c(0,0,0,0));par(mfrow=c(1,2))
am<-am_list[[1]]
#draw your first network and save the layout
lay<-draw.contagion(am,use.df=primary_tracing,day=20)
#now, if you want to use that same layout again for a different network you can just:
draw.contagion(am,am.layout=lay,use.df=secondary_tracing,day=20)
}

}
\author{
Josh Firth
}
