% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{inf_prob}
\alias{inf_prob}
\title{Samples the serial interval for given incubation period samples - gives an infection prob for a given day}
\usage{
inf_prob(
  day = NULL,
  inc_samp = NULL,
  theta = NULL,
  R = NULL,
  contactrate = NULL,
  infasym = NULL
)
}
\arguments{
\item{day}{day of simulation}

\item{inc_samp}{vector of samples from the incubation period distribution}

\item{theta}{proportion of presymptomatic transmission}

\item{R}{scaling factor}

\item{contactrate}{vector of network edge weights}

\item{infasym}{vector of weights based on whether inds are asymptomatic}
}
\value{
vector of probabilities
}
\description{
Samples the serial interval for given incubation period samples - gives an infection prob for a given day
}
\examples{



}
