% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{network_null}
\alias{network_null}
\title{Create null networks}
\usage{
network_null(
  am,
  returns = c("graph", "matrix", "edgelist"),
  null = c("edge", "deg", "latt", "clust")
)
}
\arguments{
\item{am}{association matrix network}

\item{returns}{format for output - either "graph", "matrix" or "edgelist"}

\item{null}{type of simulation to perform "edge" = randomised edges, "deg" = degrees retained, "latt" = lattice, "clust" = clustered}
}
\value{
see returns
}
\description{
Create null networks
}
\examples{

}
\author{
Josh Firth
}
