% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  net,
  num.initial.cases,
  prop.ascertain,
  cap_max_days,
  R,
  presymrate,
  delay_shape,
  delay_scale,
  prop.asym,
  quarantine,
  isolation,
  tracing,
  secondary,
  outside,
  testing,
  cap_max_tests = NULL,
  weekly = TRUE,
  s = NULL,
  test_neg
)
}
\arguments{
\item{net}{network from which to sample cases, in pairwise list format.}

\item{num.initial.cases}{integer number of initial cases}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{R}{scaling factor for infection probability}

\item{presymrate}{the proportion of cases with pre-onset transmission (must be 0<=x<=1)}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}

\item{prop.asym}{proportion of asymptomtaic cases (must be 0<=x<=1)}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{isolation}{logical whether isolation is in effect}

\item{tracing}{logical whether tracing is in effect}

\item{secondary}{logical whether secondary contact tracing is in effect}

\item{outside}{daily probability of infection from outside the network}

\item{testing}{logical whether testing is in effect}

\item{cap_max_tests}{integer - maximum number of tests}

\item{weekly}{logical - return summarised weekly cases or the raw case data}

\item{s}{seed - optional for reproducing the same output. Useful for \code{plot_network()}}
}
\value{
data.frame of cases by week, cumulative cases, and weekly isolations, quarantines and tests
}
\description{
Run a single instance of the branching process model
}
\examples{

\dontrun{
load("data-raw/am_list.RData")
haslemere <- format_network(am_list[[1]])

outbreak_model(net = haslemere, num.initial.cases = 1, prop.ascertain = 0.9, cap_max_days = 69, R = 0.8, presymrate = 0.2,
delay_shape = 1,delay_scale = 1.4,prop.asym = 0.4, quarantine = TRUE, isolation = TRUE, tracing = TRUE, secondary = TRUE,
outside = 0.001, testing = FALSE, test_neg = 0.1)}


}
\author{
Joel Hellewell

Lewis Spurgin
}
