% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{plot_network}
\alias{plot_network}
\title{Run an outbreak model and generate network plot with contagion}
\usage{
plot_network(
  am,
  am.layout = NULL,
  day,
  num.initial.cases = NULL,
  prop.ascertain = NULL,
  R = NULL,
  presymrate = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  prop.asym = NULL,
  quarantine = NULL,
  isolation = NULL,
  tracing = NULL,
  secondary = NULL,
  outside = NULL,
  testing = NULL,
  cap_max_tests = NULL,
  s = NULL
)
}
\arguments{
\item{am}{association matrix}

\item{am.layout}{(if wanting to specify a layout e.g. to keep the layout the same across comparisons)}

\item{day}{the day to do the contagion for}

\item{num.initial.cases}{integer number of initial cases}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{R}{scaling factor for infection probability}

\item{presymrate}{the proportion of cases with pre-onset transmission (must be 0<=x<=1)}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}

\item{prop.asym}{proportion of asymptomtaic cases (must be 0<=x<=1)}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{isolation}{logical whether isolation is in effect}

\item{tracing}{logical whether tracing is in effect}

\item{secondary}{logical whether secondary contact tracing is in effect}

\item{outside}{daily probability of infection from outside the network}

\item{testing}{logical whether testing is in effect}

\item{cap_max_tests}{integer - maximum number of tests}

\item{s}{seed - optional for reproducing the same output. Useful for \code{plot_network()}}
}
\value{

}
\description{
Run an outbreak model and generate network plot with contagion
}
\examples{

\dontrun{

load("data-raw/am_list.RData")
m <- am_list[[1]] #This is the association matrix for haslemere
plot_network(
am = m,
day = 20,
num.initial.cases = 1,
prop.asym = 0.4,
delay_shape =  1,
delay_scale = 1.4,
prop.ascertain = 0.8,
presymrate = 0.4,
R = 6.5,
outside = 0.001,
sensitivity = "high",
testing = "none",
isolation = FALSE,
secondary = FALSE,
tracing = FALSE,
quarantine = FALSE)
}
}
\author{
Josh Firth

Lewis Spurgin
}
