% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{plot_network_app}
\alias{plot_network_app}
\title{Plot a newtwork (shiny app version)}
\usage{
plot_network_app(
  am,
  day,
  num.initial.cases,
  prop.asym,
  delay,
  prop.ascertain,
  presymrate,
  R,
  outside,
  sensitivity,
  testing,
  cap_max_tests = NULL,
  scenario
)
}
\value{

}
\description{
Plot a newtwork (shiny app version)
}
\author{
Josh Firth

Lewis Spurgin
}
