% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Infections on one day in the network model}
\usage{
outbreak_step(
  day,
  case_data,
  net = haslemere,
  prop.asym,
  incfn,
  delayfn,
  prop.ascertain,
  presymrate,
  R,
  quarantine,
  isolation,
  tracing,
  secondary,
  outside,
  testing = FALSE,
  cap_max_tests = NULL,
  test_neg
)
}
\arguments{
\item{day}{number of days from start of the inection - must be > 0}

\item{case_data}{data.frame of cases in outbreak so far; initially generated by outbreak_setup}

\item{net}{network from which to sample cases, in pairwise list format.}

\item{prop.asym}{proportion of asymptomtaic cases (must be 0<=x<=1)}

\item{incfn}{function samples from incubation period; generated by dist_setup}

\item{delayfn}{function samples from the onset-to-hospitalisation delay; generated by dist_setup}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{presymrate}{the proportion of cases with pre-onset transmission (must be 0<=x<=1)}

\item{R}{scaling factor for infection probability}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{isolation}{logical whether isolation is in effect}

\item{tracing}{logical whether tracing is in effect}

\item{secondary}{logical whether secondary contact tracing is in effect}

\item{outside}{daily probability of infection from outside the network}

\item{testing}{logical whether testing is in effect}

\item{cap_max_tests}{integer - maximum number of tests}
}
\value{

}
\description{
Infections on one day in the network model
}
\examples{

\dontrun{
load("data-raw/am_list.RData")
haslemere <- format_network(am_list[[1]])
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(1, 1.4)
# generate initial cases
case_data <- outbreak_setup(net = haslemere, num.initial.cases = 1,incfn,delayfn,prop.asym=0.4, isolation = TRUE)
# generate next generation of cases
outbreak_step(day = 1, case_data = case_data,net = haslemere, prop.asym = 0.4, incfn = incfn, delayfn = delayfn, prop.ascertain = 1,
presymrate = 0.2, R = 0.8, quarantine = TRUE, isolation = TRUE, tracing = TRUE, secondary = TRUE, outside = 0.001, testing = FALSE,test_neg = 0.1)
}
}
\author{
Lewis Spurgin

Joel Hellewell
}
